仕様変更を伴う大規模コミットです 一時不安定になる可能性があります
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5125 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDConvexMesh.h
@@ -8,10 +8,9 @@
 #ifndef CDCONVEXMESH_H
 #define CDCONVEXMESH_H
 
-#include <SprCollision.h>
-#include <vector>
 #include <Foundation/Object.h>
-#include "CDConvex.h"
+#include <Collision/CDConvex.h>
+#include <vector>
 
 namespace Spr{;
 
@@ -79,10 +78,10 @@ public:
 	///	頂点バッファを返す。
 	virtual Vec3f* GetBase(){return &*base.begin();}	
 
-	CDFaceIf* GetFace(size_t i);
-	size_t NFace();
+	CDFaceIf* GetFace(int i);
+	int NFace();
 	Vec3f* GetVertices();
-	size_t NVertex();
+	int NVertex();
 
 	///	デスクリプタCDConvexMeshDescの読み書き	
 	virtual bool GetDesc(void *desc) const;
